+++ tests/methods/get-data.js
@@ -1,7 +1,7 @@
 describe('.getData()', function() {
     it('should get data', function(done) {
         var cropbox = new Cropbox('#plugin');
-        cropbox.getCb().addEventListener('cb:loaded', function() {
+        cropbox.getCb().addEventListener('cb:ready', function() {
             cropbox.crop();
         });
         cropbox.getCb().addEventListener('cb:cropped', function() {