+++ tests/methods/load.js
@@ -2,7 +2,7 @@ describe('.load(src)', function() {
     it('should load the image', function(done) {
         var cropbox = new Cropbox('#plugin'),
             oImg = document.createElement('img');
-        cropbox.getCb().addEventListener('cb:loaded', function() {
+        cropbox.getCb().addEventListener('cb:ready', function() {
             expect(cropbox.getImage()).to.be.equalImage(oImg);
             done();
         });